DSEGOG-267 Improve tests
- Split test classes into separate files and make some minor improvements
