+++ test/records/ingestion/test_record.py
@@ -1,0 +1,124 @@
+import pytest
+
+from operationsgateway_api.src.exceptions import (
+    ModelError,
+    RejectRecordError,
+)
+from operationsgateway_api.src.records import ingestion_validator
+from test.records.ingestion.create_test_hdf import create_test_hdf_file
+
+
+class TestRecord:
+    @pytest.mark.parametrize(
+        "shotnum, active_experiment",
+        [
+            pytest.param(
+                ["valid", "exists"],
+                ["90097341", "exists"],
+                id="all optional values present",
+            ),
+            pytest.param(
+                ["valid", "exists"],
+                ["90097341", "missing"],
+                id="only shotnum present",
+            ),
+            pytest.param(
+                ["valid", "missing"],
+                ["90097341", "exists"],
+                id="only active_experiment present",
+            ),
+        ],
+    )
+    @pytest.mark.asyncio
+    async def test_record_checks_pass(
+        self,
+        shotnum,
+        active_experiment,
+        remove_hdf_file,
+    ):
+        record_data, _, _, _ = await create_test_hdf_file(
+            shotnum=shotnum,
+            active_experiment=active_experiment,
+        )
+        record_checker = ingestion_validator.RecordChecks(record_data)
+
+        record_checker.active_area_checks()
+        record_checker.optional_metadata_checks()
+
+    @pytest.mark.parametrize(
+        "active_area, shotnum, active_experiment, test, match",
+        [
+            pytest.param(
+                ["ea1", "missing"],
+                ["valid", "exists"],
+                ["90097341", "exists"],
+                "active_area",
+                "active_area is missing",
+                id="active_area is missing",
+            ),
+            pytest.param(
+                [467, "exists"],
+                ["valid", "exists"],
+                ["90097341", "exists"],
+                "active_area",
+                "active_area has wrong datatype. Expected string",
+                id="active_area wrong datatype",
+            ),
+            pytest.param(
+                ["ea1", "exists"],
+                ["invalid", "exists"],
+                [90097341, "exists"],
+                "other",
+                "empty",
+                id="shotnum and active_experiment have wrong datatype",
+            ),
+            pytest.param(
+                ["ea1", "exists"],
+                ["invalid", "exists"],
+                ["90097341", "missing"],
+                "other",
+                "empty",
+                id="shotnum has wrong datatype active_experiment missing",
+            ),
+            pytest.param(
+                ["ea1", "exists"],
+                ["valid", "missing"],
+                [90097341, "exists"],
+                "optional",
+                "active_experiment has wrong datatype. Expected string",
+                id="active_experiment has wrong datatype shotnum missing",
+            ),
+        ],
+    )
+    @pytest.mark.asyncio
+    async def test_record_checks_fail(
+        self,
+        active_area,
+        shotnum,
+        active_experiment,
+        test,
+        match,
+        remove_hdf_file,
+    ):
+        if test == "other":
+            with pytest.raises(ModelError):
+                record_data, _, _, _ = await create_test_hdf_file(
+                    active_area=active_area,
+                    shotnum=shotnum,
+                    active_experiment=active_experiment,
+                )
+        else:
+            record_data, _, _, _ = await create_test_hdf_file(
+                active_area=active_area,
+                shotnum=shotnum,
+                active_experiment=active_experiment,
+            )
+            record_checker = ingestion_validator.RecordChecks(record_data)
+
+            with pytest.raises(RejectRecordError, match=match):
+                if test == "timestamp":
+                    record_checker.timestamp_checks()
+                elif test == "active_area":
+                    record_checker.active_area_checks()
+                elif test == "optional":
+                    record_checker.optional_metadata_checks()
